PostUploadPack: use gittest to create test repos and commits
The PostUploadPack tests are relatively old and contain lots of manual
test setup, such as creating commits with 'git commit', and resolving
the values of branches. At the time we wrote them the gitlab-test
repository was a moving target.

Nowadays, we have test helpers that create commits, and we pin the
refs of the test repo so can just hard code the commits refs are
supposed to point to in the tests.