blob: Split out function to process blobs
We're about to add another RPC which sends out blobs, where logic to
process the blobs will be largely the same as for `ListBlobs()`. To
avoid code duplication, this commit splits out most of the logic to
retrieve blobs, split them up into multiple messages if required, and
then finally to pass them back to the caller.