Re-implement Stream RPC server handling using sidechannel registry
In
https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1051#note_634147375,
we came up with a new alternative design for Stream RPC. In general,
we'd add additional RPCs (PostUploadPackStream, PackObjectsHookStream,
SSHUploadPackStream) to return a streaming token to the client once the
verifications are done and they are ready to launch git to begin
streaming the response. The RPC handler would wait for StreamRPC
connection registry.  When the client receives the streaming token. It
then dials to the listening port of Gitaly server. Gitaly server detects
tthe connection, pushes the connection to StreamRPC connection registry.
As a result, the stream RPC handler is unlocked, and continue to stream
data on pulled raw TCP connection.