repository: Allow voting on missing config
While not likely, it could be that a repository is missing a gitconfig
due to whatever reason. While we can mostly cope with it just fine, in
the context of transactional voting it would mean that voting fails
because we cannot retrieve the file's contents to compute the vote.

Fix this issue by handling the case where the config doesn't exist via a
constructed vote on an artificial value.

Changelog: fixed