helper: Remove needless condition in `DecorateError()`
The `DecorateError()` helper function explicitly checks whether the
passed error is `nil`. Given that `GrpcCode()` always returns OK in case
it's passed a `nil` error though, and that `DecorateError()` requires
the code to be Unknown in order to wrap it, this condition is needless.

Simplify the code by dropping the needless condition.