helper: Drop test of internal unused function
One of the regression tests we have asserts that the current
implementation of `ErrPreconditionFailedf()` matches what we had before
it had been refactored. While this test was a nice demonstrator back
when the interfaces were refactored, nowadays it doesn't serve much of a
purpose: we know it behaves the same, so there's not much of a need to
retain it now.

Drop the testcase.