Add a method for reserving a repository ID from the database
Praefect creates database records for a repository only after the creation
operation has been completed by the Gitalys. As Praefect will later generate
the relative paths where the Gitalys will store the replicas, it needs to
know the repository ID prior to creating the database records. This commit
adds a method for reserving a repository ID prior to creating the repository.
This allows Praefect to generate the relative path and create the repository
on the Gitalys prior to acknowledging the creation in the database. The
generated ID is consumed even if the repository creation fails. This ensures
any future creation attempts map to different disk state, so the client can
always retry a failed repository creation. For completeness, the method also
checks that no repository exists at the given virtual storage and relative path.
This is mostly to fail early. If there are two concurrent creations for the same
virtual storage and relative path, they both will get unique IDs and store the
repositories in different directories on the Gitalys. The faster one will then
win the race and create the database records with CreateRepository. The slower
will receive a RepositoryExistsError from CreateRepository and not acknowledge
the creation to the client.