Pass relative path to CreateRepository
CreateRepository will need to know the relative path of the repository
the client is trying to create in order to check whether a repository with
it already exists or not. This commit pipes the relative path to CreateRepository
to prepare for follow-up changes even though the relative path is not yet used.