Assert repository id in repository store tests
This commit changes the repository store test's to support asserting the
repository ids on the records. Since there's no code hooking up the repository
IDs yet, the only records in the test that get the repository id currently are
the `repositories` table's records due to autogeneration. Later commits will
introduce the code to link the other records.