Link records with repository ID in CreateRepository
When a repository is being created, all of the records need to be connected
via the repository ID. This commit expands CreateRepository to take in the
repository ID Praefect reserved for the repository while routing the creation
request. CreateRepository then links all of the records it creates via the
repository id. While the repository ID is not really used for anything yet, the
newly created records being linked via the ID allow us to run a migration in the
following record to just link all existing records correctly via the repository
ID. Any concurrently created records would already be linked correctly by the code
added in this commit.