git2go: rebase: Allow passing object IDs
The git2go rebase command currently only accepts a branch name and
upstream revision as inputs. While the latter is fine, requiring a
branch name is overly restrictive: all we want to have is a commit which
shall be rebased onto the upstream revision, and we may not necessarily
have a branch available. Updating references never happens inside of
gitaly-git2go either, but instead in the calling context.

Amend the command struct to also support a commit ID as input,
deprecating the old branch name parameter. For backwards compatibility
reasons, the branch name needs to stay around for at least one release.
To align types, this also introduces a new `UpstreamCommitID` field
which deprecates the previous `UpstreamRevision` field.