operations: Detect empty committer/author parameters early
We currently rely on Git commands to detect empty committer and author
signatures, which is quite fragile. Furthermore, this ties us too much
to the actual implementation given that we must execute Git in order to
get the correct error message.

Refactor the code to check username and email of both committer and
author early on to fix this. While at it, we also convert the validation
to use `errors.New()` instead of `fmt.Errorf()` and `helper.Err*f()`
instead of `status.Errorf()`.