git: Implement new `config.Set()` function
The `git.Config` interface currently only provides a function to append
multivalue config entries to the gitconfig. In the general case though,
it's more likely that the user wants to set a single entry, where any
preexisting values are overridden.

Implement a new `config.Set()` function which does exactly that.