repository: Remove feature flag for direct fetches
With the `ReplicateRepositoryDirectFetch` feature flag, we have fixed
`FetchInternalRemote()` to work in a context where called via Praefect.
Previously, we have propagated transactional information via an RPC call,
which then resulted in the remote side trying to cast votes against a
Gitaly connection, which cannot work. With the new code, we don't do an
RPC call but instead just call the relevant function directly.

This feature has been default-enabled for three weeks in production now,
and furthermore it's been backported to v13.12 and v14.0. No problems
were observed in production, and it fixed the original issue for
customers.

Given its exposure, let's remove the feature flag without taking the
intermediate step of first default-enabling it.

Changelog: fixed