repository: Remove feature flag for atomic repo creation from bundles
When not using the `--atomic` flag for git-fetch(1), then Git will
create a separate reference transaction per reference that is to be
fetched. This can be extremely expensive in case we fetch thousands or
even tens of thousands of references given that we'll have to execute
the gitaly-hooks binary each time, connect to the hooks service, cast
votes and then wait for quorum. This was fixed via the introduction of
`CreateRepositoryFromBundleAtomicFetch`, which causes us to pass that
flag.

This feature has been default-enabled for three weeks in production now,
and furthermore it's been backported to v13.12 and v14.0. No problems
were observed in production, and it fixed the performance issue for
customers.

Given its exposure, let's remove the feature flag without taking the
intermediate step of first default-enabling it.

Changelog: performance