Use a separate query for SetAuthoritativeStorage from IncrementGeneration
SetAuthoritativeStorage RPCs purpose is to set one of the replicas to the
latest generation as a way to accept data loss. Currently it uses
IncrementGeneration to do so. The behavior should be different though.
SetAuthoritativeStorage needs to modify the generation even if the targeted
replica is not on the latest generation as its purpose is to set it to the
latest one. IncrementGeneration on the other hand is used to acknowledge
mutator RPCs. It should not allow incrementing generation of replicas which
are not on the latest generation as that would effectively mean jumping over
some writes and losing them. This can currently happen if the primary changes
while there is an inflight write to the old primary. In preparation for fixing
this, this commit gives the RPC its own query so it can behave correctly even
after the behavior of IncrementGeneration is altered.