Link created assignments via the repository ID to the repository
The assignments are already being linked with the repository ID when
they are being created through CreateRepository as part of repository
creation. Assignment records can also be created when SetReplicationFactor
is called. This commit updates SetReplicationFactor to link the created
assignments records via the repository ID so we can later do a one swoop
migration to link all existing records to their repositories via the id.