gitaly: Sort the tags returned by FindAllTags RPC
The gitlab uses FindAllTags to fetch all tags and it requires
to do sorting in-memory after fetch. By this change we provide
ability to sort tags on the gitaly side by: tag name or creation
date in ascending or descending order.
The default remains the same as if no sorting requested - sort by
tag name, so it should not affect any existing clients.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3592

Changelog: added