Create repositories explicitly in Coordinator's tests
SetGeneration should not create repository records but only set the
generation of a replica. This commit creates repositories explicitly
in Coordinator's tests so we can change the behavior of SetGeneration
in a follow up commit.

StreamDirector's tests also currently create generation records with a negative
generation. Negative generations are used in some contexts to indicate a missing
record. In this case though, we should just not create the records if they
are supposed to be missing.