Remove the repository importer
The repository importer created database records for repositories that
were missing them as part of the migration to the repository specific
primaries. It only worked with the SQL elector which hasn't been usable
anymore since GitLab 14.0. This commit removes the importer from the
code base as it's no longer needed so we don't need to maintain it when
making changes to code it relies on.

The 'virtual_storages' table used by the importer can be removed in the
next release as removing them right now will fail due to the backwards
compatibility of schema tests.

Changelog: removed