Create repository records explicitly in RepositoryStore tests
RepositoryStore tests currently rely on SetGeneration to create a record
for the repository. SetGeneraton should only create records for replicas
that are being created through replication, not the repository itself.
To keep the tests working after we change SetGeneration's behavior, this
commit updates the tests to explicitly create the repository records.