Only create the replica's record in SetGeneration
SetGeneration is used to set a replica's generation after it has been
replicated to. In the past, it also created a record for the repository
if it was missing one. This has been abused in the tests to create
records for the repositories while setting a desired generation for a
replica. Repository creation logic has since expanded, and the tests
should also create repositories by using CreateRepository. SetGeneration
should never create a repository record, so this commit drops the behavior
now that tests have been updated to properly create the repository records
they expect.