proto: Remove the IsRebaseInProgress RPC
In [1], as part of [2], we rewrote `UserRebaseConfirmable`. This rewrite
does no longer use worktrees to perform the rebase. The implementation
of `IsRebaseInProgress` on the other hand uses the existence of a
worktree to determine whether a rebase is ongoing. So since the rewrite
of `UserRebaseConfirmable` the `IsRebaseInProgress` RPC started to
return false all the time.

Therefore, in [3] removed all calls of the `IsRebaseInProgress` RPC from
gitlab-org/gitlab. This was done in milestone %14.1, so now we are ready
to remove the RPC completely from the gitaly codebase.

1. 5f81d1edb (operations: Implement UserRebaseConfirmable in Go,
   2021-03-12)
2. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2876
3. https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65856

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3614
Changelog: removed