featureflag: Unconditionally enable transactional removal of repos
With a22bb54cc (coordinator: Fix repo creation/removal race for
up-to-date secondaries, 2021-07-07), we have introduced a feature flag
which enables transactions for the `RemoveRepository()` RPC. This was
required in order to plug a race when we get both a `RemoveRepository()`
and `CreateRepository()` RPC in quick succession, where the scheduled
replication job of the former RPC might kick in while we are about to
create the repository. By doing transactional voting, we avoid creating
a replication job for the deletion as long as the secondaries are
up-to-date.

This feature flag has been introduced in v14.1, but has been backported
to v14.0. Given its heightened exposure and the fact that it's been
enabled in production for some time now without any issue, this commit
removes the feature flag altogether while skipping the default-enabled
step.

Changelog: fixed