operations: Fix error when rebase in UserSquash fails
With the UserSquashWithoutWorktree flag, we're rebasing commits on top
of the start commit before doing the actual squash. But when the rebase
fails because of a conflicts, then we do not properly return a proper
Git error like we did before, when patch application on top of the start
commit failed.

Fix the issue by returning a `gitError` such that we return the correct
response.

Changelog: fixed