Add ListRef RPC to list all refs
To create an incremental backup using git bundle create we want to be
able to pass a list of ref patterns that ensure the bundle mostly only
contains the new objects since the last backup. To do this we cannot
rely on git bundle list-heads since  this would require downloading
bundles from object storage. Instead we're going to separately fetch the
entire ref list every time a backup is taken (full or incremental) then
use the previous list and the current list to calculate the required ref
patterns for git bundle create.

Changelog: added