test: Removal of the postgres build tag
As we always require a PostgreSQL DB instance to be
running during the test execution it makes no more
sense to guard SQL dependent tests with a special
build tag. The change removes the flag from the code.
As we modify files that has a lint issues we should
fix them as well otherwise the lint job will fail.

We still need praefect_sql_test CI job because it
verifies the service works properly with PgBouncer
instance between service and database.