repository: Prepare FetchRemote tests for nested repository paths
The `copyRepoWithNewRemote()` function currently assumes that the newly
created repository is always going to be stored in the root directory of
the storage. While this is the case right now, it will not be when we're
converting its callers to create "proper" repositories in the hashed
storage format.

Prepare the function to handle this correctly by creating the new
repository relative to its source repository.