Specify the revision to read a license blob from in FindLicense's Go port
FindLicense's Go port uses 'git cat-file' to read a possible license file
from the repository when the license analyzer is going through them. It
doesn't specify the treeish to read the blob from, it only gives cat-file
a ':BLOB_NAME'. This works when the repository has a worktree but doesn't
work with bare repositories as it's unclear which tree should the blob should
be read from. This commit fixes the problem by always reading the blob from
HEAD. The tests are also changed to use a bare repository as that matches
better the setup in production and catches this bug should it occur ever
again.