Choose the license match with highest confidence in FindLicense
FindLicence's Go port has a bug where it doesn't update the highest
confidence seen so far when going through the license matches. This
commit fixes the bug so the the license match with the highest
confidence is always returned.

To catch similar bugs in the future, the test is updated to use a
license file which causes multiple matches, namely 'MIT' and 'MIT-0'.