Refactor FindLicence's tests
This commit refactors FindLicence's tests to use table drive style to make
it easier to add new test cases. Additionally, the test repo is replaced with
a bare repository which each test case can set up as needed. This makes it
easier to test various cases and not to depend on any changes in the test
repository.