Return 'other' for unknown license from FindLicense
FindLicense currently returns an empty response when it fails to
identify the license of a repository. This fails some existing Rails
tests which expect to receive 'other' for unknown licenses. This commit
changes the behavior of FindLicense to return 'other' when the license
wasn't identified. Empty response is still returned if no license files
were found at all.