operations: Pass localrepo into `fetchStartRevision()`
Refactor `fetchStartRevision()` to explicitly accept a `localrepo.Repo`
as input instead of extracting it from the incoming request. This will
allow us to more easily pass in a repository that's different from what
the request has. This will be used to pass quarantined repositories into
the function instead of passing the unquarantined one.

No change in behaviour is expected from this refactoring.