operations: Support object quarantine in UserFFBranch
Implement support for object quarantine directories in UserFFBranch.
While this RPC will never create any objects but only do a fast-forward
merge of a reference, it still makes sense to support object quarantine
directories such that Rails' access checks can efficiently compute the
fact that there are no new objects.

Note that because no objects are created, we cannot really assert that
the quarantine directory is used. As a consequence, no new test is
added, but existing tests are adjusted to assert they continue to work
both with and without the feature flag.