Add Prometheus scrape timeout to configuration
This commit adds a scrape timeout to the Prometheus configuration of
both Praefect and Gitaly. The option is not used anywhere yet but will
be used to configure Praefect's SQL metrics in a follow-up commit. In
the future, the scrape timeout could be propagated further to ensure the
any long running scrapes timeout accordingly.