operations: Validate early whether UserFFBranch got a repository
We're currently relying on the fact that any of the called functions in
UserFFBranch will return an error if the request's repository is unset.
This is quite fragile and is bound to break as soon as the logic of
UserFFBranch changes.

Fix this by validating early whether the request has a repository or
not.