gitlabclient: Finalize HTTP response
When HTTP response has data in its body we need to
consume it fully before returning otherwise the connection
will remain unused because of the data that is stored in
the network buffers. Those we try to discard whatever data
is left and then close the Body of the response. To have
visibility on any problems that could appear during
that process we log errors of the body discard and closing
operations if any.