Correct the first page when paginating CommitService.CommitsBetween
Imagine we have the following set of commits:

A < B < C < D

If we call `CommitsBetween(A, D)`, we expect to get them in that order.

When using its pagination with a limit of 2, we'd expect this:

```
CommitsBetween(A, D, page_token: "", limit: 2) -> ([A, B], token1)
CommitsBetween(A, D, page_token: token1, limit: 2) -> ([C, D], token2)
CommitsBetween(A, D, page_token: token2, limit: 2) -> ([], token2)
```

Instead, we get nothing back at all.

This is because we unconditionally set `from = page_token + "~"`, which
is just `"~"`.

If we use A as the page_token, we get:

```
CommitsBetween(A, D, page_token: A, limit: 2) -> ([B, C], token1)
CommitsBetween(A, D, page_token: token1, limit: 2) -> ([D], token2)
CommitsBetween(A, D, page_token: token2, limit: 2) -> ([], ?)
```

In this case, we start at the *parent* of A, rather than A, since `~`
is unconditionally added to the page token.

The only way to get the commit as part of this RPC is to either send no
pagination params (in which case we can't limit), or to start with a
child of A (which may not exist).

We could use a separate RPC call to get A, but this is non-ideal.

If we avoid overriding `from` when a page_token is not passed, we can
provide access to A in the same RPC. As long as the final page of
commits returns the same page token (instead of "") this won't break
backward compatibility either.

Changelog: fixed