lint: Handle error returned by out.Flush
The change removes out.Flush from lint rules exclusion
and fixes all the places where it causes the issue.
It affects only gitaly-debug binary, so it has no impact
on the production code.
The change adds a couple of helper functions that now used
in places where error were not checked, but should be.