documentation: Update requirements to run the tests
Because Postgres database is required for tests to pass
we explicitly says about it in the documentation and provide
a simple instruction on how to get one.
The change also updates documentation of the glsql package
to align it with the new approach. The old info that is not
actual anymore is removed.