Introduce log field for grpc method type
This patch introduces a new log field grpc.meta.method_type which
classifies RPCs as unary, client_stream, server_stream, or bidi_stream.

This is equivalent to the grpc_type field in prometheus.

We rely on that field for alerting, as we only consider unary RPCs in
scope for the apdex. In order to more easily correlate alerts with logs
and traces, it'd be useful to have an additional field.

We can currently work around this for logs, as the msg field includes
the string unary or streaming -- but that is not true for tracing.
Having a dedicated field will help narrow down the set of samples we're
interested in.