Accept array of patterns in CreateBundleFromRefList
For lower cost streaming we should be packing as much into a frame as
possible. So instead of only taking a single pattern we should take an
array.

GRPC expects strings to be well formed UTF-8. This might not be the case
for ref names, so to allow us to write patterns for every possible ref,
we need to accept byte arrays instead.

Changelog: fixed