conflicts: Always resolve conflicts with hooks
With 83d7ef2b2 (Move mocked Manager into hook package, 2021-06-29), we
have converted the ResolveConflicts RPC to use pre-receive hooks such
that merge requests are properly updated when the RPC updates any refs.
This change was default-enabled with 2211513fe (Enable
ResolveConflictsWithHooks feature flag by default, 2021-07-01), but our
stance was that we cannot yet remove the feature flag in this release
given that the voting logic has changed. As a result, all Gitalies must
know to run the same code before we can remove the flag, otherwise their
votes may disagree.

While this reasoning makes sense, waiting for all nodes to have the code
doesn't help either because the feature flag defaults are set by Gitaly.
As a result, we still have the same issue where Gitalies may run with
different code paths and disagree even if we first default-enable the
flag. To fix this, we'd have to always inject all known feature flags at
the Praefect level such that we can atomically enable use of the flag
for all Gitalies at once.

Given that this issue exists regardless of whether we have the flag
default-enabled or removed, let's accelerate the process and remove the
flag now already. Like this, we'll always run hooks in the context of
ResolveConflicts and thus have both proper access checks and notify
Rails that refs have changed.

Changelog: fixed