operations: Allow variable flags for creating quarantined repos
Currently, all callsites of `quarantinedRepo()` pass a feature flag.
This is about to change though given that we want to remove the feature
flag for quarantined UserCreateTag, in which case one callsite will not
pass a flag anymore.

Refactor the code to take a variable number of flags: if any of the
given flags is disabled, the repository will not be quarantined. While
we currently don't care for the case where multiple flags are given,
this implicitly also covers the case where none is given.

This is a preparatory change for the subsequent commit.